+++ bindings/nostr-ffi/src/event/builder.rs
@@ -16,7 +16,7 @@ use crate::helper::unwrap_or_clone_arc;
 use crate::key::Keys;
 use crate::nips::nip01::Coordinate;
 use crate::nips::nip15::{ProductData, StallData};
-use crate::nips::nip34::GitRepositoryAnnouncement;
+use crate::nips::nip34::{GitIssue, GitRepositoryAnnouncement};
 use crate::nips::nip51::{ArticlesCuration, Bookmarks, EmojiInfo, Emojis, Interests, MuteList};
 use crate::nips::nip53::LiveEvent;
 use crate::nips::nip57::ZapRequestData;
@@ -774,6 +774,16 @@ impl EventBuilder {
     pub fn git_repository_announcement(data: GitRepositoryAnnouncement) -> Self {
         Self {
             inner: nostr::EventBuilder::git_repository_announcement(data.into()),
+        }
+    }
+
+    /// Git Issue
+    ///
+    /// <https://github.com/nostr-protocol/nips/blob/master/34.md>
+    #[uniffi::constructor]
+    pub fn git_issue(issue: GitIssue) -> Self {
+        Self {
+            inner: nostr::EventBuilder::git_issue(issue.into()),
         }
     }
 }