+++ bindings/nostr-js/src/event/builder.rs
@@ -13,7 +13,7 @@ use crate::error::{into_err, Result};
 use crate::key::{JsKeys, JsPublicKey};
 use crate::nips::nip01::JsCoordinate;
 use crate::nips::nip15::{JsProductData, JsStallData};
-use crate::nips::nip34::JsGitRepositoryAnnouncement;
+use crate::nips::nip34::{JsGitIssue, JsGitRepositoryAnnouncement};
 use crate::nips::nip51::{
     JsArticlesCuration, JsBookmarks, JsEmojiInfo, JsEmojis, JsInterests, JsMuteList,
 };
@@ -772,6 +772,16 @@ impl JsEventBuilder {
     pub fn git_repository_announcement(data: JsGitRepositoryAnnouncement) -> Self {
         Self {
             inner: EventBuilder::git_repository_announcement(data.into()),
+        }
+    }
+
+    /// Git Issue
+    ///
+    /// <https://github.com/nostr-protocol/nips/blob/master/34.md>
+    #[wasm_bindgen(js_name = gitIssue)]
+    pub fn git_issue(issue: JsGitIssue) -> Self {
+        Self {
+            inner: EventBuilder::git_issue(issue.into()),
         }
     }
 }